[ion] Missing cast

+++ ion/src/device/shared/drivers/flash.cpp
@@ -58,7 +58,7 @@ void EraseSector(int i) {
 
 void WriteMemory(uint8_t * destination, uint8_t * source, size_t length) {
   assert(SectorAtAddress((uint32_t)destination) >= 0);
-  if (AddressIsInInternalFlash(destination)) {
+  if (AddressIsInInternalFlash((uint32_t)destination)) {
     InternalFlash::WriteMemory(destination, source, length);
   } else {
     ExternalFlash::WriteMemory(destination - ExternalFlash::Config::StartAddress, source, length);
